moved fanta files to my package - Ikewolf77
