+++ src/main/java/ch/heigvd/res/chill/domain/Ikewolf77/Fanta.java
@@ -1,4 +1,4 @@
-package ch.heigvd.res.chill.domain.wasadigi;
+package ch.heigvd.res.chill.domain.Ikewolf77;
 
 import ch.heigvd.res.chill.domain.IProduct;
